Multi Image working Finally

+++ DearDiary/views.py
@@ -30,7 +30,7 @@ def Upload(request):
 def index(request):
     #return HttpResponseRedirect("base.html")
     if request.user.is_authenticated(): #or request.session.get_expiry_age()> 10):
-        request.session.set_expiry(600)
+        request.session.set_expiry(600000)
         query=request.GET.get('q')
         if query:
             return redirect('/view?q=%s' %query)
@@ -48,18 +48,26 @@ def landing(request):
        # print (request.user.is_authenticated())
         return redirect("/index")
     return render(request, "landing.html", {'form': form, 'title': title})  
-           
+          
 def create(request):
-   
   #  dogs.save()
     user = request.user
     #print (user)
     if request.method=="POST":
-     dog = Post(title=request.POST['title'],
-     description=request.POST['description'],
-     image=request.FILES.get('image'),
-     created_by_user=user,)
-     dog.save()
+        post = Post(title=request.POST['title'],
+        description=request.POST['description'],
+        created_by_user=user,) 
+        post.save()
+        
+        image=request.FILES.getlist('image')
+        for x in image:
+                 post_image = Post_image(
+                 image=x,
+                 post=post
+                    )         
+                 post_image.save()
+        #import pdb;pdb.set_trace()
+        
     return redirect("/")
         
 def error(request):
@@ -67,20 +75,15 @@ def error(request):
     
 def view(request):
     if request.user.is_authenticated(): #or request.session.get_expiry_age()> 10):
-        request.session.set_expiry(600)
+        request.session.set_expiry(600000)
         user = request.user
-    
-        # dogs=Post.objects.all() #For seeing all entries 
-        dogs= Post.objects.filter(created_by_user = user).order_by('-created_at')#[:4] #For seeing user specific entries
-        query=request.GET.get('q')
-        if query:
-            dogs=dogs.filter(
-                Q(title__icontains=query)|
-                Q(description__icontains=query)|
-                Q(created_at__icontains=query)
-        
-            ).distinct()
-        context={ 'dogs': dogs}
+        #dogs=Post.objects.all() #For seeing all entries 
+        post= Post.objects.select_related().filter(created_by_user = user).order_by('-created_at')#[:4] #For seeing user specific entries
+        # post_image=Post_image.count()
+        # posts=Post.objects.count()
+        # print post_image
+        # print posts
+        context={'posts': post}
         return render(request, 'view.html', context)
     else:
         messages.info(request, 'Session Expired')
@@ -88,47 +91,70 @@ def view(request):
     
     
 def final(request, id):
-    
     dog=Post.objects.get(id=id)
-    context={"dog": dog}
-    return render(request, 'final.html', context)
+    if dog.created_by_user==request.user:
+        if request.user.is_authenticated():
+            request.session.set_expiry(600)
+            context={"dog": dog}
+            return render(request, 'final.html', context)
+        else:
+            messages.info(request, 'Session Expired')
+            return redirect("/login")
+    else:
+        pass
+   
     
 def edit(request, id):
-    if request.user.is_authenticated(): #or request.session.get_expiry_age()> 10):
-        request.session.set_expiry(600)
-        query=request.GET.get('q')
-        if query:
-            return redirect('/view?q=%s' %query)
-        dog=Post.objects.get(id=id)
-        context={"dog": dog}
-        return render(request, 'edit.html', context)
+    dog=Post.objects.get(id=id)
+    if dog.created_by_user==request.user:
+        
+        if request.user.is_authenticated(): #or request.session.get_expiry_age()> 10):
+            request.session.set_expiry(600000)
+            query=request.GET.get('q')
+            if query:
+                return redirect('/view?q=%s' %query)
+            context={"dog": dog}
+            return render(request, 'edit.html', context)
+        else:
+            messages.info(request, 'Session Expired')
+            return redirect("/login")
     else:
-        messages.info(request, 'Session Expired')
-        return redirect("/login")
+        pass
     
 def update(request, id):
     dog=Post.objects.get(id=id)
-    dog.title=request.POST['title']
-    dog.description=request.POST['description']
-    if dog.image == None:
-        dog.image=request.FILES.get('image')
-    elif dog.image != None:
-        if request.FILES.get('image') != None:
-             dog.image=request.FILES.get('image')
-        else:
-            dog.image== dog.image
-    #dog.image=request.FILES.get('image')
-    dog.save()
-    return redirect('/view')
+    if dog.created_by_user==request.user:
+        
+        dog.title=request.POST['title']
+        dog.description=request.POST['description']
+        if dog.image == None:
+            dog.image=request.FILES.get('image')
+        elif dog.image != None:
+            if request.FILES.get('image') != None:
+                 dog.image=request.FILES.get('image')
+            else:
+                dog.image== dog.image
+        #dog.image=request.FILES.get('image')
+        dog.save()
+        return redirect('/view')
+    else:
+        pass
     
 
 def delete(request, id):
     dog=Post.objects.get(id=id)
-    dog.delete()
-    return redirect('/view')
+    if dog.created_by_user==request.user:
+     post = dog.post_image_set.all()
+     post.delete()
+     dog.delete()
+     return redirect('/view')
+     
+    else:
+        pass
 
 def delete_image(request, id):
-    Post.objects.get(id=id).image.delete(save=True)
+    post = Post_image.objects.get(id=id)
+    post.delete()
     return HttpResponseRedirect(('/view'))
     
 
@@ -165,6 +191,7 @@ def register_view(request):
     return render(request, "form.html", context)
     
 def logout_view(request):
+    
     logout(request)
     return redirect("/login")
